requirements.txt
Flask
pytesseract
scikit-image
pillow
gunicorn
Jinja2